requirements.txt
fastapi~=0.115.5
uvicorn~=0.32.1

kiss_ai_stack_core